Refactoring Move to different locations
