+++ src/Persistence/Game/ICityMapper.cs
@@ -1,0 +1,5 @@
+namespace CivOne.Persistence.Game
+{
+    public interface ICityMapper : ICityBasic, ICityStatus {}
+
+}
