+++ src/Persistence/Mapper/PlayerDtoMapper.cs
@@ -5,6 +5,9 @@ using CivOne.Advances;
 using CivOne.Civilizations;
 using CivOne.Enums;
 using CivOne.Governments;
+using CivOne.Persistence.Game;
+using CivOne.Persistence.Mapper;
+using CivOne.Persistence.Resolver;
 
 namespace CivOne.Persistence.Model
 {
