+++ src/Persistence/Mapper/SaveGameMetaDataDto.cs
