--- src/Persistence/Model/ICityMapper.cs
@@ -1,9 +1,0 @@
-namespace CivOne.Persistence.Model
-{
-	using System.Collections.Generic;
-	using System.Linq;
-    using CivOne.Tiles;
-    
-    public interface ICityMapper : ICityBasic, ICityStatus {}
-
-}
