+++ src/Persistence/Resolver/ICityDefinitionResolver.cs
@@ -3,7 +3,7 @@ using CivOne.Buildings;
 using CivOne.Enums;
 using CivOne.Wonders;
 
-namespace CivOne.Persistence.Model
+namespace CivOne.Persistence.Resolver
 {
 	public interface ICityDefinitionResolver
 	{
