+++ src/Persistence/Yaml/YamlLoadMapperDependenciesFactory.cs
@@ -5,7 +5,10 @@ using CivOne.Advances;
 using CivOne.Enums;
 using CivOne.Governments;
 using CivOne.Persistence.Factories;
+using CivOne.Persistence.Game;
+using CivOne.Persistence.Mapper;
 using CivOne.Persistence.Model;
+using CivOne.Persistence.Resolver;
 
 namespace CivOne.Persistence.Yaml
 {
