+++ xunit/src/persistence/Model/GameStateDtoMapperTest.cs
@@ -16,6 +16,10 @@ namespace CivOne.Persistence.Model
 	using CivOne.Wonders;
 	using Xunit;
 	using AdvanceId = System.UInt32;
+	using CivOne.Persistence.Mapper;
+	using CivOne.Persistence.Game;
+	using CivOne.Persistence.Resolver;
+	using CivOne.Persistence.Factories;
 
 	public class GameStateDtoMapperTest
 	{
